Rewrite job runner in Rust (#332)
## Summary

Replaces the old run.sh with a new Rust-based binary.

## Motivation

The old run.sh script had the following problems:

- Did not handle SIGTERM and pass it to the runc container correctly
- Piping logs to Vector was incredibly hacky
- Log shipping was flakey
- Spawning another instance of Vector for log shipping required extra resources

After a few attempts to fix these issues, it became increasingly difficult to do this reliably in Bash and the magical-random-symbol-toxic-spew became worse and worse.

## Other benefits

This architecutre has other added benefits:

- Very easy to test thoroughly
- Rate limiting is done as soona s possible
- We now log a custom message when lines are rate limited or truncated (which is not possible with Vector)
- The setup_* scripts can be converted to Rust eventually for better testing and reliability